requirements.txt
customtkinter
packaging
opencv-python
pillow